# JIT: extend forward sub to local fields