Optimize cases where we copy a struct and then immedately read just one field
to read from the original struct instead.