@AndyAyersMS 2022-08-10
cc @dotnet/jit-contrib 

Partially inspired by #66776. Just a small simple piece of struct copy prop.

BenchmarkDotNet=v0.13.1.1786-nightly, OS=Windows 11 (10.0.22000.856/21H2)
Intel Core i7-8700 CPU 3.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET SDK=7.0.100-rc.1.22374.1
  [Host]     : .NET 6.0.8 (6.0.822.36306), X64 RyuJIT
  Job-ZUQRZK : .NET 7.0.0 (42.42.42.42424), X64 RyuJIT
  Job-SYRZBQ : .NET 7.0.0 (42.42.42.42424), X64 RyuJIT

PowerPlanMode=00000000-0000-0000-0000-000000000000  InvocationCount=5000  IterationTime=250.0000 ms
MaxIterationCount=20  MinIterationCount=15  MinWarmupIterationCount=6
UnrollFactor=1  WarmupCount=-1

|    Method | Tool | Size |     Mean |    Error |   StdDev |   Median |      Min |      Max | Ratio | RatioSD |  Gen 0 |  Gen 1 | Allocated | Alloc Ratio |
|---------- |----- |----- |---------:|---------:|---------:|---------:|---------:|---------:|------:|--------:|-------:|-------:|----------:|------------:|
| LinqQuery | BASE |  512 | 53.49 us | 0.791 us | 0.740 us | 53.62 us | 52.50 us | 54.99 us |  1.00 |    0.00 | 5.6000 | 0.6000 |  34.33 KB |        1.00 |
| LinqQuery | DIFF |  512 | 48.59 us | 0.631 us | 0.559 us | 48.66 us | 47.16 us | 49.37 us |  0.91 |    0.02 | 5.6000 | 0.6000 |  34.33 KB |        1.00 |

Still a rough. Some things to sort through:
* various odd restrictions
* should we bail if we're making a local DNER
* should we try and handle the promoted case
* could we do some of this in the inliner
* what are the rules for when we have Layout

Some nice looking diffs, eg
```asm
;;; BEFORE

; Assembly listing for method Microsoft.CodeAnalysis.ExternalAccess.VSTypeScript.Api.VSTypeScriptCodeFixContextExtensions:IsBlocking(Microsoft.CodeAnalysis.CodeFixes.CodeFixContext):bool
; Emitting BLENDED_CODE for X64 CPU with AVX - Windows
; optimized code
; rsp based frame
; partially interruptible
; No matching PGO data
; 0 inlinees with PGO data; 1 single block inlinees; 0 inlinees without PGO data
; Final local variable assignments
;
;  V00 arg0         [V00,T00] (  3,  6   )   byref  ->  rcx         single-def
;# V01 OutArgs      [V01    ] (  1,  1   )  lclBlk ( 0) [rsp+00H]   "OutgoingArgSpace"
;  V02 tmp1         [V02,T01] (  2,  4   )  struct (56) [rsp+00H]   do-not-enreg[SF] class-hnd exact "Single-def Box Helper"
;
; Lcl frame size = 56

G_M40065_IG01:        ; gcrefRegs=00000000 {}, byrefRegs=00000000 {}, byref, nogc <-- Prolog IG
       sub      rsp, 56
       vzeroupper 
						;; size=7 bbWeight=1    PerfScore 1.25
G_M40065_IG02:        ; gcrefRegs=00000000 {}, byrefRegs=00000002 {rcx}, byref, nogc
       ; byrRegs +[rcx]
       vmovdqu  ymm0, ymmword ptr[rcx]
       vmovdqu  ymmword ptr[rsp], ymm0
       vmovdqu  xmm0, xmmword ptr [rcx+32]
       vmovdqu  xmmword ptr [rsp+20H], xmm0
       mov      rax, qword ptr [rcx+48]
       mov      qword ptr [rsp+30H], rax
						;; size=29 bbWeight=1    PerfScore 14.00
G_M40065_IG03:        ; , extend
       movzx    rax, byte  ptr [rsp+18H]
						;; size=5 bbWeight=1    PerfScore 1.00
G_M40065_IG04:        ; , epilog, nogc, extend
       add      rsp, 56
       ret      
```
becomes
```asm
;;;; AFTER

; Assembly listing for method Microsoft.CodeAnalysis.ExternalAccess.VSTypeScript.Api.VSTypeScriptCodeFixContextExtensions:IsBlocking(Microsoft.CodeAnalysis.CodeFixes.CodeFixContext):bool
; Emitting BLENDED_CODE for X64 CPU with AVX - Windows
; optimized code
; rsp based frame
; partially interruptible
; No matching PGO data
; 0 inlinees with PGO data; 1 single block inlinees; 0 inlinees without PGO data
; Final local variable assignments
;
;  V00 arg0         [V00,T00] (  3,  6   )   byref  ->  rcx         single-def
;# V01 OutArgs      [V01    ] (  1,  1   )  lclBlk ( 0) [rsp+00H]   "OutgoingArgSpace"
;* V02 tmp1         [V02    ] (  0,  0   )  struct (56) zero-ref    do-not-enreg[SF] class-hnd exact "Single-def Box Helper"
;
; Lcl frame size = 0

G_M40065_IG01:        ; gcrefRegs=00000000 {}, byrefRegs=00000000 {}, byref, nogc <-- Prolog IG
						;; size=0 bbWeight=1    PerfScore 0.00
G_M40065_IG02:        ; gcrefRegs=00000000 {}, byrefRegs=00000002 {rcx}, byref
       ; byrRegs +[rcx]
       movzx    rax, byte  ptr [rcx+24]
						;; size=4 bbWeight=1    PerfScore 2.00
G_M40065_IG03:        ; , epilog, nogc, extend
       ret   
```