@AndyAyersMS 2022-08-11
Failure is from folding
```
STMT00000 ( 0x000[E-] ... 0x00C )
               [000005] -A-X-------                         *  ASG       simd16 (copy)
               [000003] D------N---                         +--*  LCL_VAR   simd16<System.Numerics.Vector4> V02 tmp1         
               [000002] n--X-------                         \--*  OBJ       simd16<System.Numerics.Vector4>
               [000001] -----------                            \--*  ADDR      byref 
               [000000] -------N---                               \--*  LCL_VAR   simd16<System.Runtime.Intrinsics.Vector128`1[System.Single]> V00 arg0         

STMT00001 ( ??? ... ??? )
               [000007] -----------                         *  RETURN    simd16
               [000006] -------N---                         \--*  LCL_VAR   simd16<System.Numerics.Vector4> V02 tmp1 
```
to
```
    [000005]:  [000006] is only use of [000003] (V02) fold OBJ(ADDR(X)) [000002] into X [000000], with different handles
 [folding OBJ(ADDR(LCL...))] [marking V00 as multi-reg-ret] -- fwd subbing [000000]; new next stmt is

               [000007] -----------                         *  RETURN    simd16
               [000000] m------N---                         \--*  LCL_VAR   simd16<System.Runtime.Intrinsics.Vector128`1[System.Single]> V00 arg0  
```