@AndyAyersMS 2022-08-11
Probably need some vestige of the old handle check back, layout compat isn't enough.