@SingleAccretion 2022-08-11
> Failure is from folding mismatched things under return

Yes, this is one case where we still need the handle check. It will be fixed "soon" (#72887 is the prelude).

> could we do some of this in the inliner

I notice in #66776 we have a tree characteristic of struct arguments to inlinees, with an unnecessary copy made due to the `OBJ(ADDR(LCL_VAR))` wrapping by `impNormStructVal`. In principle, this wrapping can now be simply deleted, but it will lead to large diffs with some regressions mixed in due to exposing various downstream deficiencies.

> should we try and handle the promoted case

Probably not worth it? If we see a `LCL_FLD` use of a promoted local (which is currently not possible by the time forward sub runs, though that will not be true in the future), we're likely dealing with some sort of un-analyzeable reinterpretation, otherwise we'd expect the SDSU local to have been promoted and the use replaced with that of its field.

> should we bail if we're making a local DNER

Forwarding enregisterable/promoted structs, it would be the conservative option. The worry is of course that we'd pessimize all other defs/uses.

> what are the rules for when we have Layout

We only have layout on `TYP_STRUCT` local fields. `TYP_SIMD` is intentionally excluded (#69822).

Overall, I am wondering whether we should slot the `LCL_VAR -> LCL_FLD` substitution into local copy propagation instead. Seems like that'd be more powerful, and global copy propagation already has this ability.