@AndyAyersMS 2022-08-14
> Overall, I am wondering whether we should slot the `LCL_VAR -> LCL_FLD` substitution into local copy propagation instead. Seems like that'd be more powerful, and global copy propagation already has this ability.

You may be right. I'll probably give this a try.